+++ NEWS.md
@@ -1,3 +1,11 @@
+piecepackr 0.12.0
+=================
+
+New features
+------------
+
+* New ``add_opt_cache`` function allows one to add a pre-computed cache of component opt to configuration lists (#112). 
+
 piecepackr 0.11.0
 =================